cli bookmark: Revert the deprecation of explicitly required revset arguments
This reverts dd73b5a and closes 5374 as feedback in the issue has 
shown that many longtime and new users dislike it. Martin himself also
mentioned that he wants to walk back this deprecation. 

This is a perfect showcase for having a deprecation period where users can weigh in with their opinions. 

Fixes 5374